Refactor package structure and add CLI
