+++ dRFEtools/lowess/__init__.py
@@ -1,0 +1,15 @@
+"""LOWESS helpers for :mod:`dRFEtools`."""
+
+from .redundant import (
+    _cal_lowess,
+    extract_max_lowess,
+    extract_peripheral_lowess,
+    optimize_lowess_plot,
+)
+
+__all__ = [
+    "_cal_lowess",
+    "extract_max_lowess",
+    "extract_peripheral_lowess",
+    "optimize_lowess_plot",
+]
