+++ dRFEtools/metrics/__init__.py
@@ -1,0 +1,5 @@
+"""Metrics utilities for :mod:`dRFEtools`."""
+
+from .ranking import features_rank_fnc
+
+__all__ = ["features_rank_fnc"]
